+++ src/components/card-container/__tests__/card-container-test-cases.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Basic from '../examples/basic';
 import Custom from '../examples/custom';
+import Cols from '../examples/cols';
 
 const testCases = {};
 
@@ -12,6 +13,11 @@ testCases.basic = {
 testCases.noImage = {
   description: 'Full width',
   element: <Custom />
+};
+
+testCases.cols = {
+  description: 'Use cardColSize',
+  element: <Cols />
 };
 
 export { testCases };